fix(core): move getDocumentVariantType to it's own file
