+++ packages/sanity/src/core/releases/__telemetry__/releases.telemetry.ts
@@ -1,6 +1,6 @@
 import {defineEvent} from '@sanity/telemetry'
 
-import {type DocumentVariantType} from '../../util/draftUtils'
+import {type DocumentVariantType} from '../../util/getDocumentVariantType'
 
 interface VersionInfo {
   /**
